Modernize setup and development runtime environment
Thanks to Jelte for most of this.  I basically just moved the SQL
code to a migration and stood up a dev environment via make.

The downgrading of parentid to null-able was required though I do
not know why.

I added a couple of views to make interactive "select *" doable.

Added an upsert to ensure that a dev setup has a list.

I reworked the text search configurations from what was in schema.sql.
I couldn't figure out how to make the existing code work and for my
development needs it wasn't important.